+++ set.h
@@ -6,15 +6,15 @@ class Set
 {
 public:
 
-    Set(int input1, int input2,int input3,int input4, int output):
+    Set(float input1, float input2,float input3,float input4, float output):
         input1(input1), input2(input2), input3(input3), input4(input4), output(output){}
 
 
-    int input1;
-    int input2;
-    int input3;
-    int input4;
-    int output;
+    float input1;
+    float input2;
+    float input3;
+    float input4;
+    float output;
 };
 }//vhland002
 #endif // SET_H
